+++ src/shared/index.ts
@@ -78,3 +78,4 @@ export * from "./plugin-identity"
 export * from "./log-legacy-plugin-startup-warning"
 export * from "./task-system-enabled"
 export * from "./parse-tools-config"
+export { parseModelString } from "./model-string-parser"
